(config-schema-graphql-3) Make all GraphQL config types backed by serializable Snapshot
Summary: This diff begins the heavylighting of making the GraphQL
queries served by a serializiable format, instead of the definitions
levles.

It introduces a class `PipelineSnapshot` which will eventually contain
*all* the metadata about the pipeline (Dagster types, deps, etc). Right
now it only contains a `ConfigSchemaSnapshot` which contains all the
config information.

I introduced PipelineSnapshot now so we don't have to go back and change
what we are threading through. Once we get down to the config type-only
portion of the graphql schema we switch to ConfigSchemaSnapshot.

Once those were introduced this was pretty mechnical. The "checkpoint"
that this diff represents is that dagster_graphql.schema.config_types
requires no definitions to be served.

Test Plan: BK. Click around dagit.

Reviewers: alangenfeld, sashank, themissinghlink

Differential Revision: https://dagster.phacility.com/D2248