[dagit] copy over tags on resume retry
Summary:
trying to prototype a backfilling scheduler - i wanted to resume/retry a partition until it was green. To do that, we need to copy over the partition tags.

In the near future we will add a build / root id tag and maybe a resume/retry tag

Test Plan: screen shot

Reviewers: prha, sashank, bengotow

Subscribers: