(dagster-type-namespace-graphql-4) Adding CompositeSolidDefMeta and thread through to DauphinISolidDefinition
Summary: This is getting to a more challening part of the migration to
snapshot-driven-graphql. SolidDefinitions and their child definitions
are quite self-referential in the graphql schema (InputDefintiions refer
back to SolidDefintions, etc).

The goal of this diff to to get ISolidDefinition driven solely by the
snapshot classes. In order to pull it off I temporarily stash
CompositeSolidDefinition instances instead the snapshot class
(btw going to rename the meta to snapshot to avoid confusion
with pythen metaclass) so that we don't have to move over
all the dependency mapping stuff whole hog, which I will do
separately.

Depends on D2256

Test Plan: BK

Reviewers: alangenfeld, sashank, themissinghlink

Differential Revision: https://dagster.phacility.com/D2258